Distinguish element geometry builder params for parts from params for geometric elements
