+++ core/common/src/ElementProps.ts
@@ -12,7 +12,7 @@ import {
 } from "@itwin/core-geometry";
 import { CodeProps } from "./Code";
 import { EntityProps } from "./EntityProps";
-import { ElementGeometryBuilderParams } from "./geometry/ElementGeometry";
+import { ElementGeometryBuilderParams, ElementGeometryBuilderParamsForPart } from "./geometry/ElementGeometry";
 import { GeometryStreamProps } from "./geometry/GeometryStream";
 import { IModelError, IModelStatus } from "./IModelError";
 import { SubCategoryAppearance } from "./SubCategoryAppearance";
@@ -103,7 +103,7 @@ export interface GeometricElementProps extends ElementProps {
   category: Id64String;
   /** The geometry stream properties */
   geom?: GeometryStreamProps;
-  /** How to build a GeometryStream. This is not a persistent property. It may be specified as an alternative to `geom` when inserting or updating an element.
+  /** How to build the element's GeometryStream. This is used for insert and update only. It is not a persistent property. It will be undefined in the properties returned by functions that read a persistent element. It may be specified as an alternative to `geom` when inserting or updating an element.
    * @alpha
    */
   elementGeometryBuilderParams?: ElementGeometryBuilderParams;
@@ -215,10 +215,10 @@ export interface GeometricElement2dProps extends GeometricElementProps {
  */
 export interface GeometryPartProps extends ElementProps {
   geom?: GeometryStreamProps;
-  /** How to build a GeometryStream. This is not a persistent property. It may be specified as an alternative to `geom` when inserting or updating an element.
+  /** How to build the part's GeometryStream. This is used for insert and update only. It is not a persistent property. It will be undefined in the properties returned by functions that read a persistent element. It may be specified as an alternative to `geom` when inserting or updating an element.
    * @alpha
    */
-  elementGeometryBuilderParams?: ElementGeometryBuilderParams;
+  elementGeometryBuilderParams?: ElementGeometryBuilderParamsForPart;
   bbox?: LowAndHighXYZ;
 }
